// src/types/enums.ts
export enum Algorithm {
    multiplication = 'multiplication',
    maximum = 'maximum',
    minimum = 'minimum',
    average = "average",
}

export enum UnitTable {
    Force = 'stl.unittable.force',
    Stress = 'stl.unittable.stress',
    Displacement = 'stl.unittable.displacement',
    Strain = 'stl.unittable.strain',
    Area = 'stl.unittable.area',
}